+++ src/Util/include/JM/jm_portability.h
@@ -140,5 +140,36 @@ int jm_snprintf(char * str, size_t size, const char * fmt, ...);
 #define JM_VA_COPY(dest,src) dest=src
 #endif
 
+/**
+   \brief Sets the LC_NUMERIC locale for this thread only. A follow up call to
+   'jm_mtsafe_resetlocale_numeric' is needed to free the returned 'jm_locale_t'
+   object.
+  
+   \param jmloc:
+   \param value:
+     Value to set for LC_NUMERIC.
+   \return:
+     Pointer to object for reseting thread settings (locale, and
+     _configthreadlocale on Windows). NULL on failure.
+ */
+jm_locale_t* jm_mtsafe_setlocale_numeric(jm_callbacks* cb, const char* value);
+
+/**
+   \brief  Restores thread settings and locale.
+
+   This function is only allowed to be called when the current locale is set by 
+   'jm_mtsafe_setlocale_numeric', and the 'jmloc' argument must be what is
+   returned from that call. On Linux, the locale must in no way be modified since that
+   call.
+  
+   \param jmloc:
+     Return value from previous call to 'jm_mtsafe_setlocale_numeric'. Current
+     locale must be set with that function. This call will free 'jmloc', so it's
+     not allowed to be used after.
+   \return:
+     0 on success.
+ */
+int jm_mtsafe_resetlocale_numeric(jm_callbacks* cb, jm_locale_t* jmloc);
+
 /*@}*/
 #endif /* End of header file JM_PORTABILITY_H_ */