+++ src/Util/src/JM/jm_portability.c
@@ -355,3 +355,110 @@ int jm_snprintf(char * str, size_t size, const char * fmt, ...) {
     va_end (args);
     return ret;
 }
+
+struct jm_locale_t {
+#ifdef WIN32
+    char* locale_old;
+    int per_thread_locale_type_old;
+#else
+    locale_t locale_old;
+#endif
+};
+
+jm_locale_t* jm_mtsafe_setlocale_numeric(jm_callbacks* cb, const char* value) {
+
+	jm_locale_t* jmloc = (jm_locale_t*)malloc(sizeof(jm_locale_t));
+	if (!jmloc) {
+        jm_log_error(cb, module, "failed to allocate memory");
+		return NULL;
+	}
+
+#ifdef WIN32
+	{
+		char* tmp;
+
+        /* Save current thread settings. */
+        jmloc->per_thread_locale_type_old = _configthreadlocale(0);
+
+        /* Create a copy of locale, since any further calls to setlocale (e.g.
+         * from 3rd party code) will override the returned pointer. */
+        tmp = setlocale(LC_NUMERIC, NULL);
+		if (!tmp) {
+            jm_log_error(cb, module, "Failed to get current locale with 'setlocale'");
+			free(jmloc);
+			return NULL;
+		}
+        jmloc->locale_old = (char*)cb->malloc(strlen(tmp) + 1); /* + 1 for \0 */
+        strcpy(jmloc->locale_old, tmp);
+
+        /* Set LC_NUMERIC for this thread. */
+        _configthreadlocale(_ENABLE_PER_THREAD_LOCALE);
+        if (setlocale(LC_NUMERIC, value) == NULL) {
+            jm_log_error(cb, module, "Failed to call 'setlocale' for LC_NUMERIC with value: '%s'", value);
+			free(jmloc->locale_old);
+			free(jmloc);
+			return NULL;
+        }
+
+		return jmloc;
+	}
+#else /* _GNU_SOURCE */
+	{
+		locale_t nloc = NULL;
+
+		/* Retrieve old locale */
+		jmloc->locale_old = uselocale((locale_t)0);
+		if (jmloc->locale_old == (locale_t)0) {
+			jm_log_error(cb, module, "'uselocale' failed to get current locale");
+			goto err1;
+		}
+
+		/* Create new locale. */
+		nloc = newlocale(LC_NUMERIC_MASK, value, (locale_t)0);
+		if (nloc == (locale_t)0) {
+			jm_log_error(cb, module, "call failed: 'newlocale'");
+			goto err1;
+		}
+
+		/* Set new locale */
+		uselocale(nloc);
+
+		return jmloc;
+
+		/* Error handling */
+err1:
+		free(jmloc);
+		return NULL;
+	}
+#endif
+}
+
+int jm_mtsafe_resetlocale_numeric(jm_callbacks* cb, jm_locale_t* jmloc) {
+	if (jmloc == NULL) {
+		return 1; /* impl. error */
+	}
+
+#ifdef WIN32
+    setlocale(LC_NUMERIC, jmloc->locale_old);
+	cb->free(jmloc->locale_old);
+	jmloc->locale_old = NULL;
+
+    _configthreadlocale(jmloc->per_thread_locale_type_old);
+#else
+	{
+		/* Get current locale, which is expected to have been set with a previous
+		 * call to 'jm_mtsafe_setlocale_numeric'. */
+		locale_t loc = uselocale((locale_t)0);
+		if (loc == (locale_t)0) {
+			jm_log_error(cb, module, "'uselocale' failed to get current locale.");
+			return 1;
+		}
+		uselocale(jmloc->locale_old);
+
+		freelocale(loc);
+	}
+#endif
+
+	free(jmloc);
+	return 0;
+}