+++ RELEASE_NOTES.md
@@ -4,6 +4,12 @@ The release notes are typically a highlighting subset of all changes made. For f
 
 Note that version 2.1 is the first version with release notes. Please see the commit history for older versions.
 
+## 2.2.1
+
+### Minor
+- Bug fix: Correctly parse doubles when locale is not using decimal point
+- Check `variability != continuous` for non-Real variables
+
 ## 2.2
 
 ### Bug fixes